% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_response.r
\name{beta_binom_dr}
\alias{beta_binom_dr}
\title{Beta binomial dose-response model}
\usage{
beta_binom_dr(alpha, beta, dose)
}
\arguments{
\item{alpha}{shape parameter of the standard beta distribution}

\item{beta}{shape paramter of the standard beta distribution}

\item{dose}{vector of doses}
}
\description{
Beta binomial dose-response model
}
