% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_response.r
\name{betapois_dr}
\alias{betapois_dr}
\title{beta-poisson dose-response model}
\usage{
betapois_dr(alpha, beta, dose)
}
\arguments{
\item{alpha}{shape parameter of the standard beta distribution}

\item{beta}{shape paramter of the standard beta distribution}

\item{dose}{vector of doses}
}
\description{
beta-poisson dose-response model
}
