% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_response.r
\name{betapois_dr_approx}
\alias{betapois_dr_approx}
\title{approximation to the beta-poisson dose-response model. derived by furomoto and Mickey 1967}
\usage{
betapois_dr_approx(alpha, beta, dose)
}
\arguments{
\item{alpha}{shape parameter of the standard beta distribution}

\item{beta}{shape paramter of the standard beta distribution}

\item{dose}{vector of doses}
}
\description{
approximation to the beta-poisson dose-response model. derived by furomoto and Mickey 1967
}
