% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{betapois_N50_approx}
\alias{betapois_N50_approx}
\title{median infectious dose of the approximate beta-poisson dose-response model}
\usage{
betapois_N50_approx(alpha, beta)
}
\arguments{
\item{alpha}{shape parameter of the standard beta distribution}

\item{beta}{shape paramter of the standard beta distribution}
}
\description{
median infectious dose of the approximate beta-poisson dose-response model
}
