% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_qmra.r
\name{conc_qmra}
\alias{conc_qmra}
\title{Function to carry out Monte Carlo based simulation
for QMRA procedure}
\usage{
conc_qmra(efficacy, microbe.influent, dilution.exposure)
}
\arguments{
\item{efficacy}{either a named list with elements \code{min} and \code{max} specifying the
minimum and maximum respectively of the log10 redction expected or a sinlge numeric value.}

\item{microbe.influent}{\code{distribution} and \code{params}
specifying the distribution as a character and the parameters as a named list respectively
of the dilution distribution.}

\item{dilution.exposure}{a named list eith elements
specifying the parameters of a cumulative distribution from which to draw the dilution value: 
\code{min}, \code{max}, and the vectors \code{x} and \code{p}
where \code{min} < \code{x} < \code{max} and 0 <= \code{p} <= 1}
}
\value{
a simulated microbe value at exposure site
}
\description{
Function to carry out Monte Carlo based simulation
for QMRA procedure
}
