% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative.r
\name{dcumulative}
\alias{dcumulative}
\title{CDF of the cumulative distribution}
\usage{
dcumulative(y, x, p, min = NULL, max = NULL)
}
\arguments{
\item{y}{vector of quantiles}

\item{x}{supplied vector of quantiles (x_1, ..., x_n)}

\item{p}{supplied vector of probabilities (p_1, ..., p_n) corresponding to elements of \code{x}}

\item{min}{(optional) numeric minimum value of \code{x}, if supplied this is taken as the first
element of \code{x} and the p_0 element of \code{p} is set to be 0}

\item{max}{(optional) numeric maximum value of \code{x}, if supplied this is taken as the first
element of \code{x} and the p_(n+1) element of \code{p} is set to be 1}
}
\description{
CDF of the cumulative distribution
}
