% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_qmra.r
\name{dr.mod}
\alias{dr.mod}
\title{function to choose which dose-response model for use in qmra model fit}
\usage{
dr.mod(model, params, dose)
}
\arguments{
\item{model}{character specifying assumed type of dose-response model}

\item{params}{named vector of dose-response parameters}

\item{dose}{vector of dose values}
}
\description{
function to choose which dose-response model for use in qmra model fit
}
