% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_response.r
\name{expon_dr}
\alias{expon_dr}
\title{exponential dose-response model}
\usage{
expon_dr(r, dose)
}
\arguments{
\item{r}{pathogen-host survival probability}

\item{dose}{vector of doses}
}
\description{
exponential dose-response model
}
