% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{expon_N50}
\alias{expon_N50}
\title{median infectious dose of the exponential dose-response model}
\usage{
expon_N50(r)
}
\arguments{
\item{r}{pathogen-host survival probability}
}
\description{
median infectious dose of the exponential dose-response model
}
