% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_qmra.r
\name{fit.qmra}
\alias{fit.qmra}
\title{Function to fit dose-response model
fitting function for the various dose-response relationships 
parameter estimates obtained by minimization of the deviance \code{qmra_dev}}
\usage{
fit.qmra(data, params, model, upper, lower, method)
}
\arguments{
\item{data}{data frame of dose-respone study. Must contain named columns \code{positive},
number of positive subjects in at a specific dose;
\code{total}, total number of subjects; and \code{dose}, dose given.}

\item{params}{named vector of dose-response parameters}

\item{model}{character specifying assumed type of dose-response model}

\item{upper}{vector of upper limits of the parameters to be estimated}

\item{lower}{vector of lower limits of parameters to be estimated}

\item{method}{to carry out minimisation}
}
\description{
Function to fit dose-response model
fitting function for the various dose-response relationships 
parameter estimates obtained by minimization of the deviance \code{qmra_dev}
}
