% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_response.r
\name{fractional_pois_dr}
\alias{fractional_pois_dr}
\title{dose-response model as per Messner et al 2014}
\usage{
fractional_pois_dr(p, mu, dose)
}
\arguments{
\item{mu}{parameter of the dose response model, set to 1 in strict case. Defined as
the mean aggregate size and is a function of alpha: -alpha/((1 - alpha)*ln(1 - alpha))}

\item{dose}{vector of doses}

\item{P}{parameter of the dose-response model}
}
\description{
dose-response model as per Messner et al 2014
}
