% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{get.sim.value}
\alias{get.sim.value}
\title{Function to simulate from appropriate distribution (internal to qmra)}
\usage{
get.sim.value(x, nsim)
}
\arguments{
\item{x}{a list with named components specifying the distribution andparameter values of that distribution}

\item{nsim}{number of simulations to carry out}
}
\description{
Function to simulate from appropriate distribution (internal to qmra)
}
