% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_response.r
\name{Ghyper_dr}
\alias{Ghyper_dr}
\title{Gaus hypergeometric form dose-response model}
\usage{
Ghyper_dr(alpha, beta, k, dose)
}
\arguments{
\item{alpha}{shape parameter of the standard beta distribution}

\item{beta}{shape paramter of the standard beta distribution}

\item{k}{overdispersion parameter}

\item{dose}{vector of doses}
}
\description{
Gaus hypergeometric form dose-response model
}
