% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard.r
\name{hazard}
\alias{hazard}
\title{hazard model for probability of illness given infection}
\usage{
hazard(gamma, eta, dose, increase)
}
\arguments{
\item{gamma}{shape parameter}

\item{eta}{location parameter}

\item{dose}{vector of doses}

\item{increase}{logical is the model an increasing hazard model or a decreasing model}
}
\description{
hazard model for probability of illness given infection
}
