% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illness.R
\name{illness.qmra}
\alias{illness.qmra}
\title{Function that takes a vector of individual dose values and calculates
for a supplied dose-response function with supplied parameter values
a list of with components 1) the probability of infection and two
indicator variables indicating if individual was 2) infected
and 3) conditional on 2) ill.}
\usage{
illness.qmra(dose, model, params, p.ill)
}
\arguments{
\item{dose}{vector of dose values}

\item{model}{character specifying assumed type of dose-response model}

\item{params}{named vector of dose-response parameters}

\item{p.ill}{numeric values giving the probability of a person getting ill given they are infected}
}
\description{
Function that takes a vector of individual dose values and calculates
for a supplied dose-response function with supplied parameter values
a list of with components 1) the probability of infection and two
indicator variables indicating if individual was 2) infected
and 3) conditional on 2) ill.
}
