% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_qmra.r
\name{ingested_qmra}
\alias{ingested_qmra}
\title{Function to simulate ingested volume for QMRA models}
\usage{
ingested_qmra(duration, vol.rate)
}
\arguments{
\item{duration}{a named list with elements \code{distribution} (character), \code{min} (numeric),
\code{mode} (numeric), and \code{max} (numeric) specifying the assumed duration distribution and
associated parameters}

\item{vol.rate}{a named list with elements \code{distribution} (character), \code{min} (numeric),
\code{mode} (numeric), and \code{max} (numeric) specifying the assumed volume rate distribution and
associated parameters}
}
\description{
Function to simulate ingested volume for QMRA models
}
