% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_response.r
\name{loglogistic_dr}
\alias{loglogistic_dr}
\title{log logistic dose response model}
\usage{
loglogistic_dr(q1, q2, dose)
}
\arguments{
\item{q1}{parameter of the log-logistic dose response model}

\item{q2}{parameter of the log-logistic dose response model}

\item{dose}{vector of doses}
}
\description{
log logistic dose response model
}
