% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_response.r
\name{logprobit_dr}
\alias{logprobit_dr}
\title{log probit dose response model}
\usage{
logprobit_dr(q1, q2, dose)
}
\arguments{
\item{q1}{parameter of the log-probit dose response model}

\item{q2}{parameter of the log-probit dose response model}

\item{dose}{vector of doses}
}
\description{
log probit dose response model
}
