% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_qmra.r
\name{mc_qmra}
\alias{mc_qmra}
\title{Function to carry out Monte Carlo based simulations using both
\link{conc_qmra} and \link{ingested_qmra}}
\usage{
mc_qmra(nsim, efficacy, microbe.influent, dilution.exposure, duration,
  vol.rate)
}
\arguments{
\item{nsim}{either a numeric value specifying the number of Monte Carlo simulations to carry out for
both concentration and exposure, else a named vector of length two specifying the number of simulations seperatly.
for the \code{nsim.conc} and \code{nsim.exposure}.}

\item{efficacy}{either a named list with elements \code{min} and \code{max} specifying the
minimum and maximum respectively of the log10 redction expected or a sinlge numeric value.}

\item{microbe.influent}{\code{distribution} and \code{params}
specifying the distribution as a character and the parameters as a named list respectively
of the dilution distribution.}

\item{dilution.exposure}{a named list eith elements
specifying the parameters of a cumulative distribution from which to draw the dilution value: 
\code{min}, \code{max}, and the vectors \code{x} and \code{p}
where \code{min} < \code{x} < \code{max} and 0 <= \code{p} <= 1}

\item{duration}{a named list with elements \code{distribution} (character), \code{min} (numeric),
\code{mode} (numeric), and \code{max} (numeric) specifying the assumed duration distribution and
associated parameters}

\item{vol.rate}{a named list with elements \code{distribution} (character), \code{min} (numeric),
\code{mode} (numeric), and \code{max} (numeric) specifying the assumed volume rate distribution and
associated parameters}
}
\description{
Function to carry out Monte Carlo based simulations using both
\link{conc_qmra} and \link{ingested_qmra}
}
