% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure.r
\name{mpn_neg_ll}
\alias{mpn_neg_ll}
\title{returns the negative log-likelihood of the binomial distribution for MPN}
\usage{
mpn_neg_ll(lambda, ni, pi, volume)
}
\arguments{
\item{lambda}{the mean density of the poisson distribution}

\item{ni}{number of replicates in the set}

\item{pi}{vector of positive replicates}

\item{volume}{volume per assay replicate}
}
\description{
returns the negative log-likelihood of the binomial distribution for MPN
}
