% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illness.R
\name{multi.illness.qmra}
\alias{multi.illness.qmra}
\title{function to run \code{illness.qmra} for multiple dose-response curves and return a list for each
dose-response model with named elements containing the sum (for each run) of each vector of indicator variables
indicating if individual was 1) infected and 2) conditional on 1) ill.}
\usage{
multi.illness.qmra(dose, params.list, p.ill, mod.vec)
}
\arguments{
\item{dose}{vector of dose values}

\item{params.list}{a list of named parameter vectors for each dose-response model given in \link{mod.vec}}

\item{p.ill}{numeric values giving the probability of a person getting ill given they are infected}

\item{mod.vec}{a character vector of dose-response model names}
}
\description{
function to run \code{illness.qmra} for multiple dose-response curves and return a list for each
dose-response model with named elements containing the sum (for each run) of each vector of indicator variables
indicating if individual was 1) infected and 2) conditional on 1) ill.
}
