% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illness.R
\name{multi.qmra.dose}
\alias{multi.qmra.dose}
\title{function to run \code{illness.qmra} for multiple dose-response curves and return a data frame}
\usage{
multi.qmra.dose(dose, params.list, mod.vec)
}
\arguments{
\item{dose}{vector of dose values}
}
\description{
function to run \code{illness.qmra} for multiple dose-response curves and return a data frame
}
