% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{my.rug}
\alias{my.rug}
\title{Internal rug plot function}
\usage{
my.rug(x, ...)
}
\arguments{
\item{x}{vector of numeric values to show in a rug plot}

\item{...}{other arguments to pass into plot}
}
\description{
Internal rug plot function
}
