% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_response.r
\name{od_expon_dr}
\alias{od_expon_dr}
\title{overdispersed exponential dose-response model}
\usage{
od_expon_dr(r, k, dose)
}
\arguments{
\item{r}{pathogen-host survival probability}

\item{k}{overdispersion parameter}

\item{dose}{vector of doses}
}
\description{
overdispersed exponential dose-response model
}
