% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_qmra.r
\name{pars.qmra}
\alias{pars.qmra}
\title{function to get fitted parameter estimates of a dose-response model}
\usage{
pars.qmra(fit)
}
\arguments{
\item{qmra}{a \code{qmra} fitted object}
}
\description{
function to get fitted parameter estimates of a dose-response model
}
