% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_response.r
\name{phi}
\alias{phi}
\title{function for the log-probit dose response model}
\usage{
phi(arg)
}
\arguments{
\item{arg}{argument of this function}
}
\description{
function for the log-probit dose response model
}
