% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure.r
\name{poisson_gr}
\alias{poisson_gr}
\title{first derivative of the poisson log-likelihood}
\usage{
poisson_gr(lambda, count, volume)
}
\arguments{
\item{lambda}{the mean density of the poisson distribution}

\item{count}{a vector of pathogen counts}

\item{volume}{a vector of volumes within which the counts \code{count} were recorded}
}
\description{
first derivative of the poisson log-likelihood
}
