% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure.r
\name{poisson_neg_ll}
\alias{poisson_neg_ll}
\title{functions for the concentration/ duration distributions
returns the negative log-likelihood of a Poisson distribution}
\usage{
poisson_neg_ll(lambda, count, volume)
}
\arguments{
\item{lambda}{the mean density of the poisson distribution}

\item{count}{a vector of pathogen counts}

\item{volume}{a vector of volumes within which the counts \code{count} were recorded}
}
\description{
functions for the concentration/ duration distributions
returns the negative log-likelihood of a Poisson distribution
}
