% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_qmra.r
\docType{class}
\name{qmra}
\alias{qmra}
\alias{qmra-class}
\title{The qmra class}
\description{
The qmra class
}
\section{Slots}{

\describe{\item{\code{data}:}{a data frame which must contain the following named elements positive, total, and dose to be used with \code{fit.pcod}}
\item{\code{model}:}{a character of the chosen model based on arguments given to \code{fit.qmra}, set by user}
\item{\code{fit}:}{contains the model output}}
}

