% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_qmra.r
\name{qmra_dev}
\alias{qmra_dev}
\title{deviance statistic to be minimised for ML estimate of dose-response parameters}
\usage{
qmra_dev(params, data, model)
}
\arguments{
\item{params}{named vector of dose-response parameters}

\item{data}{data frame of dose-respone study. Must contain named columns \code{positive},
number of positive subjects in at a specific dose;
\code{total}, total number of subjects; and \code{dose}, dose given.}

\item{model}{character specifying assumed type of dose-response model}
}
\description{
deviance statistic to be minimised for ML estimate of dose-response parameters
}
