% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_qmra.r
\name{qmra_ll_sup}
\alias{qmra_ll_sup}
\title{maximum possible log-likelihood l_sup}
\usage{
qmra_ll_sup(data)
}
\arguments{
\item{data}{data frame of dose-respone study. Must contain named columns \code{positive},
number of positive subjects in at a specific dose;
\code{total}, total number of subjects; and \code{dose}, dose given.}

\item{params}{named vector of dose-response parameters}

\item{model}{character specifying assumed type of dose-response model}
}
\description{
maximum possible log-likelihood l_sup
}
