% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hockey.r
\name{rhockey}
\alias{rhockey}
\title{random number generator Hockey stick distribution}
\usage{
rhockey(n, a, b, c, d, h1, h2)
}
\arguments{
\item{n}{number of rvs to generate}

\item{a}{minimum value}

\item{b}{central measure, if traditional Hockey stick distribution then 50% of mass either side of b
(i.e., Median)}

\item{c}{right tie of distribution}

\item{d}{maximum value}

\item{h1}{height of distribution at b, only required if h2 is NULL}

\item{h2}{height of distribution at c, only required if h1 is NULL, if
both h1 and h2 are supplied then h2 will be ignored.}
}
\description{
random number generator Hockey stick distribution
}
