% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_response.r
\name{s_binom_dr}
\alias{s_binom_dr}
\title{simple binomial dose-response model}
\usage{
s_binom_dr(r, dose)
}
\arguments{
\item{r}{pathogen-host survival probability}

\item{dose}{vector of doses}
}
\description{
simple binomial dose-response model
}
