% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_response.r
\name{s_thres_dr}
\alias{s_thres_dr}
\title{simple threshold model, extension of the exponential model.}
\usage{
s_thres_dr(kmin, r, dose)
}
\arguments{
\item{kmin}{parameter of the simple threshold model. Minimum number of surviving organisms}

\item{r}{pathogen-host survival probability}

\item{dose}{vector of doses}
}
\description{
simple threshold model, extension of the exponential model.
}
