% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{show.qmra}
\alias{show.qmra}
\title{Function to plot either a fitted dose-response model
or the raw data}
\usage{
show.qmra(x)
}
\arguments{
\item{x}{a \code{qmra} fitted object or a data frame with named columns dose, positive, total}
}
\description{
Function to plot either a fitted dose-response model
or the raw data
}
