% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{tornado.qmra}
\alias{tornado.qmra}
\title{Tornado (sensitivity) qmra plot}
\usage{
tornado.qmra(mc_qmra, illness.qmra)
}
\arguments{
\item{mc_qmra}{output from \code{mc_qmra}}

\item{illness.qmra}{output from \code{illness.qmra}}
}
\description{
Tornado (sensitivity) qmra plot
}
