% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_response.r
\name{weibull_dr}
\alias{weibull_dr}
\title{weibull dose-response model}
\usage{
weibull_dr(q1, q2, dose)
}
\arguments{
\item{q1}{parameter of the Weibull dose response model}

\item{q2}{parameter of the Weibull dose response model}

\item{dose}{vector of doses}
}
\description{
weibull dose-response model
}
