% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hockey.r
\name{Xp}
\alias{Xp}
\title{Finding the "hinge" point (c parameter) for the hockey stick distribution at the pth percentile}
\usage{
Xp(b, d, h1, q = 0.05, s = 0.45)
}
\arguments{
\item{b}{central measure, if traditional Hockey stick distribution then 50% of mass either side of b
(i.e., Median)}

\item{d}{maximum value}

\item{h1}{height of distribution at b, only required if h2 is NULL}

\item{q}{= 1 - p (percentile) by default is 1 - 0.95 = 0.05}

\item{s}{= p - 0.5 (for Median b) by default is 0.95 - 0.5 = 0.45}
}
\description{
Finding the "hinge" point (c parameter) for the hockey stick distribution at the pth percentile
}
